Moving firebase config
