+++ src/components/Form/Form.tsx
@@ -1,50 +1,22 @@
-import React, { useEffect, useState } from 'react'
+import React, { useState } from 'react'
 import './Form.css'
 import { UploadButton } from '../UploadButton/UploadButton'
-import { Challenge, Course, Track } from '../../model/ChallengeConfiguration'
+import { Challenge, Course } from '../../model/ChallengeConfiguration'
 import { GPXFile } from '../../model/gpx-file'
 import { title } from 'process'
-import firebase from 'firebase'
+import { firebaseDB } from '../../model/firebase-config'
+import { useTracks } from '../../model/useFiles'
 
 interface FormProps {
   addFiles: (file: any) => void
   clearFiles: () => void
   files: Array<GPXFile>
 }
-var firebaseConfig = {
-  apiKey: 'AIzaSyCiT4b8_qyC2ZxF95aRAdE8j4Kej3Dt9kk',
-  authDomain: 'asyncracing-d5302.firebaseapp.com',
-  databaseURL: 'https://asyncracing-d5302-default-rtdb.firebaseio.com',
-  projectId: 'asyncracing-d5302',
-  storageBucket: 'asyncracing-d5302.appspot.com',
-  messagingSenderId: '468547346614',
-  appId: '1:468547346614:web:0a5a388705b91b5b088617',
-  measurementId: 'G-199TR40228',
-}
 
-firebase.initializeApp(firebaseConfig)
-
-firebase.analytics()
-
-const db = firebase.database()
-
-export const Form = ({ addFiles, clearFiles, files }: FormProps) => {
+export const Form = ({ files, addFiles, clearFiles }: FormProps) => {
+  const tracks = useTracks(files)
   const [metadata, setMetadata] = useState<Challenge['metadata']>({})
   const [course, setCourse] = useState<Course>({ start: null, finish: null })
-  const [tracks, setTracks] = useState<Array<Track>>([])
-
-  useEffect(() => {
-    ;(async () => {
-      const newTracksPromises = files.map((file: GPXFile) => {
-        return file.tracks()
-      })
-      const newTracks = await Promise.all(newTracksPromises)
-
-      //do this
-      setTracks(newTracks.flat())
-    })()
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [files.length])
 
   return (
     <div className="wrapper">
@@ -54,7 +26,7 @@ export const Form = ({ addFiles, clearFiles, files }: FormProps) => {
       <form
         onSubmit={async (e) => {
           e.preventDefault()
-          const challengeRef = db.ref('challenges')
+          const challengeRef = firebaseDB.ref('challenges')
           const challenge = await challengeRef.get()
           const newChallengeRef = challengeRef.push({
             course: {
